python refactoring wip
