+++ python/VERSION.txt
