+++ python/pdal/PyPipeline.hpp
@@ -3,8 +3,9 @@
 #include <pdal/util/FileUtils.hpp>
 #include <pdal/plang/Array.hpp>
 
+#include "PipelineExecutor.hpp"
+
 #include <string>
-#include <Python.h>
 #undef toupper
 #undef tolower
 #undef isspace
@@ -19,20 +20,19 @@ namespace libpdalpython
 {
 //     typedef std::shared_ptr<pdal::plang::Array> PArray;
     typedef pdal::plang::Array* PArray;
+
 class Pipeline {
 public:
     Pipeline(std::string const& xml);
     ~Pipeline(){};
 
-    void execute();
-    inline const char* getJSON() const { return m_json.c_str(); }
-    inline const char* getSchema() const { return m_schema.c_str(); }
+    int64_t execute();
+    inline const char* getJSON() const { return m_executor.getJSON(); }
+    inline const char* getSchema() const { return m_executor.getSchema(); }
     std::vector<PArray> getArrays() const;
 
 private:
-    std::string m_json;
-    std::string m_schema;
-    pdal::PipelineManager m_manager; // no progress reporting
+    pdal::executor::Pipeline m_executor;
 
 };
