+++ src/util/FileUtils.cpp
@@ -82,7 +82,7 @@ istream *openFile(string const& filename, bool asBinary)
         return &cin;
 
     if (!FileUtils::fileExists(name))
-        return NULL;
+        return nullptr;
 
     ios::openmode mode = ios::in;
     if (asBinary)
@@ -92,7 +92,7 @@ istream *openFile(string const& filename, bool asBinary)
     if (!ifs->good())
     {
         delete ifs;
-        return NULL;
+        return nullptr;
     }
     return ifs;
 }
@@ -111,7 +111,7 @@ ostream *createFile(string const& name, bool asBinary)
     if (!ofs->good())
     {
         delete ofs;
-        return NULL;
+        return nullptr;
     }
     return ofs;
 }
